% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countypops}
\alias{countypops}
\title{U.S. Census Bureau population estimates by county (1990-2017)}
\format{A data frame with five variables: \code{year}, \code{state} (including "District of Columbia"),
  \code{hispanic_origin} ("Hispanic" or "Non-Hispanic"),  \code{race} (racial groups
  change over time: NA (1970-1980); "White", "Black", "American Indian or Alaska Native", 
  "Asian or Pacific Islander" (1981-1999); "White", "Black", "American Indian or Alaska Native",
  "Asian", "Native Hawaiian or Other Pacific Islander", "Two or More Races" (2000-2017)), 
  \code{sex} ("Male" or "Female"), \code{age_group} (groups before 1990 are different), \code{pop}. \code{n}.}
\usage{
countypops
}
\description{
U.S. Census Bureau population estimates by county (1990-2017)
}
\keyword{datasets}
