% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_pops}
\alias{us_pops}
\title{US Census Bureau national population data, 1980-2018}
\format{A data frame (specifically a tbl_df) with 12,168 observations and 6 variables:
\describe{
  \item{year}{Year of population estimate.}
  \item{origin}{"Hispanic" or "Non-Hispanic". (Data not available for 1980-1989)}
  \item{race}{For 1980-1989, the races are "White", "Black", and "Other Race".
  For 1990-1999, the races are "White", "Black", "American Indian or Alaska Native", 
  and "Asian or Pacific Islander". For 2000-2018, the races are "White", "Black", 
  "American Indian or Alaska Native", "Asian", "Native Hawaiian or Other Pacific Islander", 
  and "Two or More Races", which counts people who were recorded as belonging to more than 
  one race (e.g., an individual recorded as "Black" and "Asian" is counted in 
  "Two Or More Races" but is not counted in "Black" or "Asian").}
  \item{sex}{"Male" or "Female".}
  \item{age}{Ages are grouped in the following order: "0-4", "5-9", "10-14", "15-19", 
  "20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59", "60-64",
  "65-69", "70-74", "75-79", "80-84", and "85+".}
  \item{pop}{Population estimate.}
}}
\usage{
us_pops
}
\description{
US Census Bureau national population data, 1980-2018
}
\keyword{datasets}
