% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_pops}
\alias{us_pops}
\title{US population data, 1900-2018}
\format{A data frame (specifically a tbl_df) with 118 observations and 2 variables: 
        \code{year} and \code{pop}, the US population on July 1st of the given year 
        as estimated by the US Census Bureau. Two exceptions are 1970 and 1980 where 
        \code{pop} is the estimated US population on April 1st. For years before 1970,
        \code{pop} is rounded to the nearest thousandth. From 1900 to 1949, the US 
        population estimate includes the first 48 states and District of Columbia.
        From 1950 to 2018, the US population includes the first 48 states, Alaska
        (granted statehood in 1959), Hawaii (granted statehood in 1959), and 
        District of Columbia. The US territory Puerto Rico is not included any 
        US population estimate.}
\usage{
us_pops
}
\description{
US population data, 1900-2018
}
\keyword{datasets}
